remove useless metrics, update readme and install, reorganize scripts
