+++ README.md
@@ -3,6 +3,5 @@ Dépôt pour le projet final de l'évaluation du cours Mise en production (ENSAE
 
 Exécuter `install.sh` pour configurer l'environnement et le projet.
 
-Le code du projet est pour l'instant dans un notebook (`notebooks/projet_time_series.ipynb`). Les métriques sont dans `src/metrics.py`
+Les données sont stockées dans le stockage externe S3 au lien suivant : https://minio.lab.sspcloud.fr/laurinemir/diffusion
 
-Les données sont stockées dans le stockage externe S3 au lien suivant : https://minio.lab.sspcloud.fr/laurinemir/taxi_data
