--- patches/softdtw.patch
@@ -1,13 +1,0 @@
-diff --git a/setup.py b/setup.py
-index 159b396..16cf2fa 100644
---- a/setup.py
-+++ b/setup.py
-@@ -30,7 +30,7 @@ setup(
-     package_data={},
-     install_requires=[
-         'numba',
--        'numpy',
-+        'numpy<2.2,>=1.24',
-         'scipy>=1.2.0',
-     ],
-     classifiers=[
